refactor(GameResultDto) : dto 내부 도메인 객체 제거
Owner -> String 으로 변경
